guard the triton import for the group-gemm
